requirements.txt
flask
flask_restful
Flask-SQLAlchemy
pymysql